Update Course.java
Added 4 new global variables because they could not be defined inside calculateForPeriod method
Changed some methods from public to private as they only serve a purpose inside the class.
Created a findModule method and awardMark with the purpose of awarding marks to a module using the module ID. This helps in creating a more real life friendly app.

Created a getDetails private method with the only purpose of avoiding duplicating codes in printing the lambda expressions.
Reset the values of the global variables at the top of the calculateForPeriod method, as the variables were keep increasing their values which every call.